api/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
google-genai>=1.0.0
pydantic>=2.5.0
